removed global state from the password ids maps

+++ internal/api/password/delete_password/delete_password.go
@@ -11,19 +11,25 @@ import (
 	show "pass_web/internal/api/show"
 )
 
-func Handler(w http.ResponseWriter, r *http.Request) {
-	id := mux.Vars(r)["id"]
+func Handler(ps *show.PasswordIdStore) http.HandlerFunc {
+	return func(w http.ResponseWriter, r *http.Request) {
+		id := mux.Vars(r)["id"]
 
-	passwordItem := show.PasswordsID[id]
-	passwordFile := passwordItem.Password
-	passwordPath := filepath.Join(passwordItem.Path, passwordFile)
-	log.Println(passwordPath)
+		passwordItem, ok := ps.Get(id)
+		if !ok {
+			http.Error(w, "Password inaccesible", http.StatusBadRequest)
+			return
+		}
+		passwordFile := passwordItem.Password
+		passwordPath := filepath.Join(passwordItem.Path, passwordFile)
+		log.Println(passwordPath)
 
-	err := os.Remove(passwordPath)
-	if err != nil {
-		log.Println(err.Error())
-		http.Error(w, "Failed to delete password", http.StatusInternalServerError)
-		return
+		err := os.Remove(passwordPath)
+		if err != nil {
+			log.Println(err.Error())
+			http.Error(w, "Failed to delete password", http.StatusInternalServerError)
+			return
+		}
+		w.WriteHeader(http.StatusOK)
 	}
-	w.WriteHeader(http.StatusOK)
 }
